Field structure changed

+++ Node.cs
@@ -3,10 +3,10 @@
 public class Node
 {
     public State State;
-    public int Depth;
+    public byte Depth;
     public Node? Previous;
 
-    public Node(State state, int depth, Node? previous)
+    public Node(State state, byte depth, Node? previous)
     {
         State = state;
         Depth = depth;
@@ -18,7 +18,7 @@ public class Node
         foreach (State state in State.GetChildren())
         {
             if (Previous is not null && Previous.State.Equals(state)) continue;
-            yield return new Node(state, Depth + 1, this);
+            yield return new Node(state, (byte)(Depth + 1), this);
         }
     }
